Add a button that asks for a "Grid Size"
This button on the top of the screen sends the user a popup asking for
the number of squares per side for the new grid. Once entered, the
existing grid is removed and a new grid is generated in the same total
space as before. The maximum number of squares is set to 100.